requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
bcrypt==3.2.0
python-dotenv==0.19.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
requests==2.31.0